backend/pyproject.toml
[tool.poetry]
name = "idp"
version = "0.1.0"
description = ""
authors = ["cerob <cerob@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.4"
djangorestframework = "^3.14.0"
Markdown = "^3.4.4"
djangorestframework-simplejwt = "^5.3.0"
numpy = "^1.25.2"
drf-yasg = "^1.21.7"
python-dotenv = "^1.0.0"
setuptools = "^69.0.2"
django-cors-headers = "^4.3.1"
scikit-learn = "^1.3.2"
pandas = "^2.1.4"
matplotlib = "^3.8.2"
plotly = "^5.18.0"
gensim = "^4.3.2"
transformers = "^4.36.0"
# torch = "^2.1.2" # CUDA enabled on AMD architecture
torch = { url = "https://download.pytorch.org/whl/cpu/torch-2.2.1%2Bcpu-cp310-cp310-linux_x86_64.whl" } # CUDA disabled; CPU-only on AMD architecture
sentence-transformers = "^2.3.1"
nltk = "^3.8.1"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"

[tool.poetry.group.dev.dependencies]
yapf = "^0.40.1"
pylint = "^2.17.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
